+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `formService.searchAPINSWLiquorLicences()`
+- `formService.getAPINSWLiquorLicence()`
+
 ## [8.0.0] - 2023-12-12
 
 ### Changed